feat(integration): DB 서브플로우에 이상 센서 데이터 알림(Notification) 발행 연동

+++ src/main/java/com/nhnacademy/processing/dto/alert/AlertEvent.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 import java.util.List;
 
 public record AlertEvent (
-        Long roomId,
+        Integer roomId,
         String alertType,
         String alertTitle,
